refactor, separating service and util files
